+++ next-docs/public/examples/caption/__tests__/index.test.tsx
@@ -9,6 +9,8 @@ import Example from '../Example';
 import Colors from '../Colors';
 import SemanticTags from '../SemanticTags';
 import TextAlign from '../TextAlign';
+import Sizes from '../Sizes';
+import FontWeight from '../FontWeight';
 
 const renderWithTheme = (component: JSX.Element) => (
   <ThemeProvider theme={moonDesignLight}>{component}</ThemeProvider>
@@ -21,6 +23,14 @@ describe('Caption', () => {
   });
   it('renders colors', () => {
     const testRenderer = renderer.create(renderWithTheme(<Colors />));
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+  it('renders sizes', () => {
+    const testRenderer = renderer.create(renderWithTheme(<Sizes />));
+    expect(testRenderer.toJSON()).toMatchSnapshot();
+  });
+  it('renders font weights', () => {
+    const testRenderer = renderer.create(renderWithTheme(<FontWeight />));
     expect(testRenderer.toJSON()).toMatchSnapshot();
   });
   it('renders text-align', () => {