+++ next-docs/public/examples/checkbox/Checked.tsx
@@ -1,11 +1,15 @@
 import { Checkbox } from '@heathmont/moon-core';
+import { useState } from 'react';
 
-const Example = () => (
-  <Checkbox
-    checked
-    label="I agree to receive bonus &amp; marketing emails."
-    readOnly
-  />
-);
+const Example = () => {
+  const [isChecked, setIsChecked] = useState(true);
+  return (
+    <Checkbox
+      checked={isChecked}
+      label="I agree to receive bonus &amp; marketing emails."
+      onClick={() => setIsChecked(!isChecked)}
+    />
+  );
+};
 
 export default Example;