+++ next-docs/public/examples/dialog/footer/ButtonMd.tsx
@@ -1,14 +1,17 @@
 import React from 'react';
-import { Footer } from '@heathmont/moon-core';
-import { Button } from '@heathmont/moon-components';
+import { Footer, Button } from '@heathmont/moon-core';
 
 const Example = () => {
   return (
     <div className="flex flex-col">
       <div className="flex justify-around items-center w-96 mb-4">
+        <Footer primButton={<Button fullWidth>Button</Button>} />
+      </div>
+      <div className="flex justify-around items-center w-96 mb-4">
         <Footer
-          primButton={
-            <Button fullWidth variant="primary" size="xsmall">
+          primButton={<Button fullWidth>Button</Button>}
+          secButton={
+            <Button fullWidth variant="secondary">
               Button
             </Button>
           }
@@ -16,32 +19,14 @@ const Example = () => {
       </div>
       <div className="flex justify-around items-center w-96 mb-4">
         <Footer
-          primButton={
-            <Button fullWidth variant="primary" size="xsmall">
-              Button
-            </Button>
-          }
+          primButton={<Button fullWidth>Button</Button>}
           secButton={
-            <Button fullWidth variant="tertiary" size="xsmall">
-              Button
-            </Button>
-          }
-        />
-      </div>
-      <div className="flex justify-around items-center w-96 mb-4">
-        <Footer
-          primButton={
-            <Button fullWidth variant="primary" size="xsmall">
-              Button
-            </Button>
-          }
-          secButton={
-            <Button fullWidth variant="tertiary" size="xsmall">
+            <Button fullWidth variant="secondary">
               Button
             </Button>
           }
           tertButton={
-            <Button fullWidth variant="secondary" size="xsmall">
+            <Button fullWidth variant="tertiary">
               Button
             </Button>
           }
@@ -49,13 +34,9 @@ const Example = () => {
       </div>
       <div className="flex justify-around items-center w-96 mb-4">
         <Footer
-          primButton={
-            <Button fullWidth variant="primary" size="xsmall">
-              Button
-            </Button>
-          }
+          primButton={<Button fullWidth>Button</Button>}
           tertButton={
-            <Button fullWidth variant="secondary" size="xsmall">
+            <Button fullWidth variant="tertiary">
               Button
             </Button>
           }
@@ -64,7 +45,7 @@ const Example = () => {
       <div className="flex justify-around items-center w-96 mb-4">
         <Footer
           secButton={
-            <Button fullWidth variant="tertiary" size="xsmall">
+            <Button fullWidth variant="secondary">
               Button
             </Button>
           }
@@ -73,12 +54,12 @@ const Example = () => {
       <div className="flex justify-around items-center w-96 mb-4">
         <Footer
           secButton={
-            <Button fullWidth variant="tertiary" size="xsmall">
+            <Button fullWidth variant="secondary">
               Button
             </Button>
           }
           tertButton={
-            <Button fullWidth variant="secondary" size="xsmall">
+            <Button fullWidth variant="tertiary">
               Button
             </Button>
           }