+++ next-docs/public/examples/select/multiSelect/Default.tsx
@@ -1,7 +1,11 @@
 import React from 'react';
-import { MultiSelect } from '@heathmont/moon-core';
-import { ListItem, Footer, Checkbox } from '@heathmont/moon-core';
-import { Button } from '@heathmont/moon-components';
+import {
+  ListItem,
+  Footer,
+  Checkbox,
+  Button,
+  MultiSelect,
+} from '@heathmont/moon-core';
 
 const SubText: React.FC = () => (
   <>Lorem Ipsum is simply dummy text of the printing and typesetting industry</>
@@ -62,18 +66,14 @@ const Example = () => (
       search={<div className="p-4">Search component</div>}
       footer={
         <Footer
-          primButton={
-            <Button variant="primary" size="xsmall">
-              Button
-            </Button>
-          }
+          primButton={<Button size="small">Button</Button>}
           secButton={
-            <Button variant="tertiary" size="xsmall">
+            <Button variant="secondary" size="small">
               Button
             </Button>
           }
           tertButton={
-            <Button variant="secondary" size="xsmall">
+            <Button variant="tertiary" size="small">
               Button
             </Button>
           }